+++ wxc_sdk/as_api.py
@@ -51,8 +51,8 @@ CALLING_DATA_TIMEOUT_PROTECTION = False
 
 
 __all__ = ['AsAccessCodesApi', 'AsAgentCallerIdApi', 'AsAnnouncementApi', 'AsAnnouncementsRepositoryApi',
-           'AsApiChild', 'AsAppServicesApi', 'AsAttachmentActionsApi', 'AsAutoAttendantApi', 'AsBargeApi',
-           'AsCQPolicyApi', 'AsCallInterceptApi', 'AsCallParkApi', 'AsCallPickupApi', 'AsCallQueueApi',
+           'AsApiChild', 'AsAppServicesApi', 'AsAttachmentActionsApi', 'AsAuthorizationsApi', 'AsAutoAttendantApi',
+           'AsBargeApi', 'AsCQPolicyApi', 'AsCallInterceptApi', 'AsCallParkApi', 'AsCallPickupApi', 'AsCallQueueApi',
            'AsCallRecordingApi', 'AsCallWaitingApi', 'AsCallerIdApi', 'AsCallingBehaviorApi',
            'AsCallparkExtensionApi', 'AsCallsApi', 'AsDetailedCDRApi', 'AsDeviceConfigurationsApi',
            'AsDeviceSettingsJobsApi', 'AsDevicesApi', 'AsDialPlanApi', 'AsDndApi', 'AsEventsApi',
@@ -178,6 +178,80 @@ class AsAttachmentActionsApi(AsApiChild, base='attachment/actions'):
         url = self.ep(f'{action_id}')
         data = await super().get(url=url)
         return AttachmentAction.model_validate(data)
+
+
+class AsAuthorizationsApi(AsApiChild, base='authorizations'):
+    """
+    Authorizations are user grants to applications to act on the user's behalf. Authorizations are how Integrations get
+    authorized with specific access scopes in the oAuth client life-cycle. Integrations and some of the Webex service
+    portals, like developer.webex.com, are all oAuth clients, each with their unique clientId.
+
+    Your application receives an API access token and a refresh token through the oAuth process. The access token is
+    used to call Webex APIs for which the user authorized the scopes. Access tokens expire fairly frequently,
+    while refresh tokens (when being regularly used) will be refreshed to last forever (see Using the Refresh Token
+    for details).
+
+    In this API an authorization is synonymous with an API access token.
+
+    To provide admins with fine-grained token management control, you use the /authorizations API with the DELETE HTTP
+    method to revoke access and refresh tokens.
+
+    Deleting a refresh token will revoke all associated access tokens as well. Deleting an access token will revoke the
+    developers ability to call the APIs with it. Webex subsystems may cache the validity of the token for a short while
+    longer after the authorization was deleted.
+
+    Admins can revoke user authorizations for users in their organization. When an admin deletes their own token,
+    the clientId used to authorize the request must match the clientId used to generate the token.
+
+    To use the authorizations API in an Integration the scopes must
+    include: identity:tokens_write, identity:tokens_read.
+    """
+
+    async def list(self, person_id: str = None, person_email: str = None) -> list[Authorization]:
+        """
+        Lists all authorizations for a user. Either personId or personEmail must be provided. This API does not support
+        pagination.
+
+        :param person_id: List authorizations for this user id.
+        :param person_email: List authorizations for this user email.
+        :return: List of Authorizations
+        """
+        params = {to_camel(k): v
+                  for k, v in locals().items()
+                  if k not in {'self'} and v is not None}
+        if frozenset(params) not in {frozenset({'personId'}),
+                                     frozenset({'personEmail'})}:
+            raise ValueError(
+                'Invalid parameter combination: exactly one of person_id or person_email has to be present.')
+        url = self.ep()
+        data = await self.get(url, params=params)
+        return TypeAdapter(list[Authorization]).validate_python(data['items'])
+
+    async def delete(self, authorization_id: str = None, client_id: str = None, org_id: str = None):
+        """
+        Deletes an authorization, by authorization ID or client ID and org ID.
+
+        Specify the authorization Id in the authorizationId parameter in the URI which was listed in the list
+        resource. The client_id parameter can be combined with org_id.
+
+        :param authorization_id: The unique identifier for the authorization
+        :type authorization_id: str
+        :param client_id: The unique oAuth client id.
+        :type client_id: str
+        :param org_id: The ID of the organization. If no orgId is specified, use orgId from the OAuth token.
+        :type org_id: str
+        """
+        params = {to_camel(k): v
+                  for k, v in locals().items()
+                  if k not in {'self'} and v is not None}
+        if frozenset(params) not in {frozenset({'authorizationId'}),
+                                     frozenset({'clientId'}),
+                                     frozenset({'clientId', 'orgId'})}:
+            raise ValueError(
+                'Invalid parameter combination: exactly one of authorization_id or client_id has to be present '
+                'and org_id can only be combined with client_id. ')
+        url = self.ep()
+        await super().delete(url, params=params)
 
 
 class AsDetailedCDRApi(AsApiChild, base='devices'):
@@ -14578,6 +14652,8 @@ class AsWebexSimpleApi:
 
     #: Attachment actions API :class:`AsAttachmentActionsApi`
     attachment_actions: AsAttachmentActionsApi
+    #: Authorizations API :class:`AsAuthorizationsApi`
+    authorizations: AsAuthorizationsApi
     #: CDR API :class:`AsDetailedCDRApi`
     cdr: AsDetailedCDRApi
     #: device configurations API :class:`AsDeviceConfigurationsApi`
@@ -14648,6 +14724,7 @@ class AsWebexSimpleApi:
 
         session = AsRestSession(tokens=tokens, concurrent_requests=concurrent_requests, retry_429=retry_429)
         self.attachment_actions = AsAttachmentActionsApi(session=session)
+        self.authorizations = AsAuthorizationsApi(session=session)
         self.cdr = AsDetailedCDRApi(session=session)
         self.device_configurations = AsDeviceConfigurationsApi(session=session)
         self.devices = AsDevicesApi(session=session)